Moved directory functions to seperate files.
